First alpha version of grayscale support
How to use:
1.  Create your figure normally, do whatever you're feeling like.
2.  Set the global flag rcParams['gray'].
3.  Render your figure.

There is also support for rendering only parts of the figure in grayscale
by setting for Artists .set_gray(True).  The flag applies to the Artist it
is set in and all artists which are drawn by this Artist.

The Artist-flag feature is entirely untested, but I believe it works nearly
out of the box.

How it works:
*   A global flag rcParams['gray']
*   Local flags in all Artists: ._gray, .set_gray(), .get_gray().
*   Decorator take_gray_into_account for Artist.draw().
    This sets the global flag for the time the Artist is drawn.
*   colors.ColorConverter respects the grayscale flag.

Because all color conversions go back to colors.colorConverter somewhen,
this, as long as this holds, ensures that the grayscale flag is taken
into account.

There were some changes in collections.py and image.py to make sure the
colors are recalculated on draw time.  They are virtually conflicting
with the caching optimisations they work against.

Also, effort has been put into completely refactoring colors.ColorConverter.

All the draw() methods have been decorated with take_gray_into_account.