Enabled symbolic links for gray mode
Symbolic links are objects where the gray mode shall be taken
from.  If an Artist has an symbolic link gray mode, its gray
mode will be that of the object linked.

The symbolic link is set via the :func:`Artist.set_gray`, with
another Artist as argument (instead of the boolena truth value).

There is currently no check for infinite link loops.