Colorbars now inherit the gray mode
:func:`figure.Figure.colorbar` not establishes a gray link
from the colorbar axes (cax) to the axes the colobar belongs
to (ax).

Thus the gray mode of the colorbar now is always that of the
corresponding Axes.

Note: The :class:`colobar.Colorbar` constructor doesn't have
access to the corresponding original Axes anymore, only to the
colorbar Axes.  Thus implementing the feature in
:class:`figure.Figure` is the most general way of doing it.