Modified inheritance of gray mode for colorbars
Old behaviour: On call of Figure.colorbar(), the cax links to the ax.

New behaviour: On call of Colorbar.__init__(), the cax links to the
mappable.

Rationale:

1) Some calls from axes_grid1 did not use Figure.colorbar().

2) It is not safe to make use of the *ax* argument in Figure.colorbar(),
since it might be determined by a "current axes" query in plt.colorbar(),
and I do not trust in that this always meets our specification to it in
this case.

3) It is more sane and more fundamental to link to the mappable, since the
colorbar itself also refers to a mappable and not to an axes.

If the old behaviour, i.e. linkage to the Axes, is needed, a manual relink
will do.

Nasty: axes_grid1 duplicates a lot of the Colorbar code without inheritance.
So maintenance will be painful.